Integrate plugin suppression with the services approach
