+++ includes/options/class-amp-options-menu.php
@@ -106,18 +106,11 @@ class AMP_Options_Menu {
 			]
 		);
 
-		if ( count( PluginSuppression::get_suppressible_plugins() ) > 0 ) {
-			add_settings_field(
-				Option::SUPPRESSED_PLUGINS,
-				__( 'Plugin Suppression', 'amp' ),
-				[ $this, 'render_suppressed_plugins' ],
-				AMP_Options_Manager::OPTION_NAME,
-				'general',
-				[
-					'class' => 'amp-suppressed-plugins',
-				]
-			);
-		}
+		/**
+		 * This fire when settings fields for the AMP Options menu need to be
+		 * registered.
+		 */
+		do_action( 'amp_options_menu_items' );
 
 		$submenus = [
 			new AMP_Analytics_Options_Submenu( AMP_Options_Manager::OPTION_NAME ),
@@ -440,227 +433,6 @@ class AMP_Options_Menu {
 				</li>
 			<?php endforeach; ?>
 		</ul>
-		<?php
-	}
-
-	/**
-	 * Render suppressed plugins.
-	 *
-	 * @since 1.6
-	 */
-	public function render_suppressed_plugins() {
-		?>
-		<fieldset>
-			<h4 class="title hidden"><?php esc_html_e( 'Suppressed Plugins', 'amp' ); ?></h4>
-			<p>
-				<?php esc_html_e( 'When a plugin emits invalid markup that causes an AMP validation error, one option is to review the invalid markup and allow it to be removed. Another option is to suppress the plugin from doing anything when rendering AMP pages. What follows is the list of active plugins with any causing validation errors being highlighted. If a plugin is emitting invalid markup that is causing validation errors and this plugin is not necessary on the AMP version of the page, it can be suppressed.', 'amp' ); ?>
-			</p>
-
-			<style>
-			#suppressed-plugins-table {
-				margin-top: 20px;
-			}
-			#suppressed-plugins-table th {
-				font-weight: 400;
-			}
-			#suppressed-plugins-table th,
-			#suppressed-plugins-table td {
-				padding: 8px 10px;
-			}
-			#suppressed-plugins-table .column-status {
-				width: 120px;
-			}
-			#suppressed-plugins-table .column-status > select {
-				width: 100%;
-			}
-			#suppressed-plugins-table .column-plugin {
-				width: 45%;
-			}
-			#suppressed-plugins-table .column-plugin .plugin-author-uri {
-				margin-top: 0;
-			}
-			#suppressed-plugins-table .column-details {
-				width: 50%;
-			}
-			#suppressed-plugins-table tbody th,
-			#suppressed-plugins-table tbody td {
-				vertical-align: top;
-			}
-
-			#suppressed-plugins-table details > ul {
-				margin-left: 30px;
-				margin-top: 0.5em;
-				margin-bottom: 1em;
-				list-style-type: disc;
-			}
-			#suppressed-plugins-table summary {
-				user-select: none;
-				cursor: pointer;
-				line-height: 30px; /* To match .wp-core-ui select */
-			}
-			li.error-removed {
-				color: <?php echo esc_html( Icon::valid()->get_color() ); ?>;
-			}
-			li.error-kept {
-				color: <?php echo esc_html( Icon::invalid()->get_color() ); ?>;
-			}
-			li.error-unreviewed > a {
-				background-color: #fef7f1;
-				text-shadow: #d54e21 0 0 1px;
-			}
-			@media screen and (max-width: 782px) {
-				#suppressed-plugins-table summary {
-					line-height: 40px; /* To match .wp-core-ui select */
-				}
-				#suppressed-plugins-table {
-					display: table;
-				}
-				#suppressed-plugins-table th,
-				#suppressed-plugins-table td {
-					display: table-cell;
-				}
-			}
-			</style>
-
-			<?php
-			$suppressed_plugins = AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS );
-			$plugins            = array_intersect_key( // Note that wp_array_slice_assoc() doesn't preserve sort order.
-				PluginRegistry::get_plugins( true ),
-				array_fill_keys( PluginSuppression::get_suppressible_plugins(), true )
-			);
-
-			$errors_by_sources = AMP_Validated_URL_Post_Type::get_recent_validation_errors_by_source();
-			$select_options    = [
-				'0' => __( 'Active', 'amp' ),
-				'1' => __( 'Suppressed', 'amp' ),
-			];
-			?>
-			<table id="suppressed-plugins-table" class="wp-list-table widefat fixed striped">
-				<thead>
-					<th class="column-status" scope="col"><?php esc_html_e( 'Status', 'amp' ); ?></th>
-					<th class="column-plugin" scope="col"><?php esc_html_e( 'Plugin', 'amp' ); ?></th>
-					<th class="column-details" scope="col"><?php esc_html_e( 'Details', 'amp' ); ?></th>
-				</thead>
-				<tbody>
-					<?php foreach ( $plugins as $plugin_slug => $plugin ) : ?>
-						<?php
-						$is_suppressed = array_key_exists( $plugin_slug, $suppressed_plugins );
-						?>
-						<tr>
-							<th class="column-status" scope="row">
-								<select name="<?php echo esc_attr( sprintf( '%s[%s][%s]', AMP_Options_Manager::OPTION_NAME, Option::SUPPRESSED_PLUGINS, $plugin_slug ) ); ?>">
-									<?php foreach ( $select_options as $value => $text ) : ?>
-										<option value="<?php echo esc_attr( $value ); ?>" <?php selected( (string) $is_suppressed, $value ); ?>>
-											<?php echo esc_html( $text ); ?>
-										</option>
-									<?php endforeach; ?>
-								</select>
-							</th>
-							<td class="column-plugin">
-								<?php $needs_details = ( ! empty( $plugin['Description'] ) || ! empty( $plugin['Author'] ) ); ?>
-
-								<?php if ( $needs_details ) : ?>
-									<details><summary>
-								<?php endif; ?>
-
-								<?php if ( ! empty( $plugin['PluginURI'] ) ) : ?>
-									<a href="<?php echo esc_url( $plugin['PluginURI'] ); ?>" target="_blank">
-								<?php endif; ?>
-								<strong><?php echo esc_html( $plugin['Name'] ); ?></strong>
-								<?php if ( ! empty( $plugin['PluginURI'] ) ) : ?>
-									</a>
-								<?php endif; ?>
-
-								<?php if ( $needs_details ) : ?>
-									</summary>
-								<?php endif; ?>
-
-								<?php if ( ! empty( $plugin['Author'] ) ) : ?>
-									<p class="plugin-author-uri">
-										<small>
-										<?php
-										if ( ! empty( $plugin['AuthorURI'] ) ) {
-											$author = sprintf( '<a href="%s" target="_blank">%s</a>', esc_url( $plugin['AuthorURI'] ), esc_html( $plugin['Author'] ) );
-										} else {
-											$author = esc_html( $plugin['Author'] );
-										}
-										/* translators: %s is author name */
-										echo wp_kses_post( sprintf( __( 'By %s', 'amp' ), $author ) );
-										?>
-										</small>
-									</p>
-								<?php endif; ?>
-
-								<?php if ( ! empty( $plugin['Description'] ) ) : ?>
-									<div class="plugin-description">
-										<?php echo wp_kses_post( wpautop( $plugin['Description'] ) ); ?>
-									</div>
-								<?php endif; ?>
-
-								<?php if ( $needs_details ) : ?>
-									</details>
-								<?php endif; ?>
-							</td>
-							<td class="column-details">
-								<?php if ( $is_suppressed ) : ?>
-									<p>
-									<?php if ( isset( $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_TIMESTAMP ] ) ) : ?>
-										<?php
-										printf(
-											/* translators: %s is the date at which suppression occurred */
-											esc_html__( 'Since %s.', 'amp' ),
-											sprintf(
-												'<time datetime="%s">%s</time>',
-												esc_attr( gmdate( 'c', $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_TIMESTAMP ] ) ),
-												esc_html( date_i18n( get_option( 'date_format' ), $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_TIMESTAMP ] ) )
-											)
-										);
-										?>
-									<?php endif; ?>
-
-									<?php if ( isset( $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_USERNAME ] ) ) : ?>
-										<?php
-										$user = get_user_by( 'slug', $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_USERNAME ] );
-										if ( $user instanceof WP_User ) {
-											if ( wp_get_current_user()->user_nicename === $user->user_nicename ) {
-												esc_html_e( 'Done by you.', 'amp' );
-											} else {
-												/* translators: %s is a user */
-												echo esc_html( sprintf( __( 'Done by %s.', 'amp' ), $user->display_name ) );
-											}
-										} else {
-											/* translators: %s is a user */
-											echo esc_html( sprintf( __( 'Done by %s.', 'amp' ), $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_USERNAME ] ) );
-										}
-										?>
-									<?php endif; ?>
-
-									<?php if ( version_compare( $suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_LAST_VERSION ], $plugin['Version'], '!=' ) ) : ?>
-										<?php if ( $plugin['Version'] ) : ?>
-											<?php
-											echo esc_html(
-												sprintf(
-													/* translators: %1: version at which suppressed, %2: current version */
-													__( 'Now updated to version %1$s since suppressed at %2$s.', 'amp' ),
-													$plugin['Version'],
-													$suppressed_plugins[ $plugin_slug ][ Option::SUPPRESSED_PLUGINS_LAST_VERSION ]
-												)
-											);
-											?>
-										<?php else : ?>
-											<?php esc_html_e( 'Plugin updated since last suppressed.', 'amp' ); ?>
-										<?php endif; ?>
-									<?php endif; ?>
-									</p>
-								<?php elseif ( ! $is_suppressed && ! empty( $errors_by_sources['plugin'][ $plugin_slug ] ) ) : ?>
-									<?php self::render_validation_error_details( $errors_by_sources['plugin'][ $plugin_slug ] ); ?>
-								<?php endif ?>
-							</td>
-						</tr>
-					<?php endforeach; ?>
-				</tbody>
-			</table>
-		</fieldset>
 		<?php
 	}
